[refactor] AppState ui.utils 디렉토리로 이동
